database: Execute feed and document connection
The database is feed with the test results after all test runs
completed. For the database connection the configuration of a global
Jenkins credential is required. Add documentation how to setup this
credential. To be more flexible, hand in hostname and port of the
database as a parameter. This change could be further improved by
defining the database table but depends on a change of the feed
database global variable in ci-rt Jenkins library.

The based for this change were two patches original developed by
Martin Kaistra.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>