Jenkinsfile: Load CI-RT Jenkins lib dynamic
When a different version of the CI-RT library is required, a new
Jenkinsfile needs to be specified with the proper version at the
moment. When loading the library dynamic and hand in the version of
the library as a parameter, this single Jenkinsfile can be used for
different versions of CI-RT library. The default version is master
branch of CI-RT library.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>